V2/requirements.txt
Pillow
Pyro5
serpent
requests
python-dotenv

